requirements.txt
accelerate==0.30.1
aiohttp==3.9.5
aiosignal==1.3.1
anaconda-anon-usage==0.4.4
annotated-types==0.7.0
anthropic==0.26.1
anyio==4.3.0
archspec==0.2.3
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
bitsandbytes==0.42.0
boilerpy3==1.0.7
boltons==23.0.0
Brotli==1.0.9
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
conda==24.4.0
conda-content-trust==0.2.0
conda-libmamba-solver==24.1.0
conda-package-handling==2.2.0
conda_package_streaming==0.9.0
cryptography==42.0.5
dataclasses-json==0.6.6
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
docarray==0.40.0
filelock==3.14.0
frozenlist==1.4.1
fsspec==2024.5.0
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
haystack-ai==2.1.2
haystack-bm25==1.0.2
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.1
idna==3.7
Jinja2==3.1.4
jiter==0.4.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.1
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.1
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
langsmith==0.1.63
lazy-imports==0.3.1
libmambapy==1.5.8
llama-index==0.10.39
llama-index-agent-openai==0.2.5
llama-index-cli==0.1.12
llama-index-core==0.10.39.post1
llama-index-embeddings-openai==0.1.10
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.21
llama-index-multi-modal-llms-openai==0.1.6
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.23
llama-index-readers-llama-parse==0.1.4
llama-parse==0.4.3
llamaindex-py-client==0.1.19
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
menuinst==2.0.2
mistralai==0.2.0
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
openai==1.30.3
orjson==3.10.3
packaging==23.2
pandas==2.2.2
pillow==10.3.0
pip==24.0
platformdirs==3.10.0
pluggy==1.0.0
posthog==3.5.0
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycosat==0.6.6
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pyparsing==3.1.2
pypdf==4.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
rich==13.7.1
rsa==4.9
ruamel.yaml==0.17.21
safetensors==0.4.3
scipy==1.13.1
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
striprtf==0.0.26
sympy==1.12
tenacity==8.3.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.2
transformers==4.40.0
truststore==0.8.0
types-requests==2.32.0.20240523
typing_extensions==4.12.0
typing-inspect==0.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.1.0
wheel==0.43.0
wrapt==1.16.0
yarl==1.9.4
zstandard==0.22.0
